@@ -1,6 +1,11 @@
-module.exports = {
-    plugins: [
-      require('tailwindcss'),
-      require('autoprefixer'),
-    ],
+// export const plugins = {
+//     tailwindcss: {},
+//     autoprefixer: {},
+// };
+export default {
+    plugins: {
+      tailwindcss: {},
+      autoprefixer: {},
+    },
   };
+  
