@@ -1,14 +1,14 @@
+import flowbite from "flowbite-react/tailwind";
+
 /** @type {import('tailwindcss').Config} */
-const flowbite = require('flowbite-react/tailwind');
-module.exports = {
-    content: ["./src/**/*.{js,jsx,ts,tsx}",
-      flowbite.content()
-    ],
-    theme: {
-      extend: {},
-    },
-    plugins: [
-      flowbite.plugin()
-    ],
-  };
-  
+export default {
+  content: [
+    "./index.html", 
+    "./src/**/*.{js,ts,jsx,tsx}", 
+    flowbite.content(),
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [flowbite.plugin()],
+};
